+++ packages/expo-router/src/native-tabs/NativeTabTrigger.tsx
@@ -4,7 +4,7 @@ import { useCallback, type ReactElement, type ReactNode } from 'react';
 import { StyleSheet } from 'react-native';
 
 import { useIsPreview } from '../link/preview/PreviewRouteContext';
-import { useNavigation, useRoute } from '../react-navigation/native';
+import { useNavigation } from '../react-navigation/native';
 import { useFocusEffect } from '../useFocusEffect';
 import { filterAllowedChildrenElements, isChildOfType } from '../utils/children';
 import {
@@ -55,7 +55,6 @@ import { appendIconOptions } from './utils/optionsIconConverter';
  * ```
  */
 function NativeTabTriggerImpl(props: NativeTabTriggerProps) {
-  const route = useRoute();
   const navigation = useNavigation();
   const isInPreview = useIsPreview();
 
@@ -65,11 +64,14 @@ function NativeTabTriggerImpl(props: NativeTabTriggerProps) {
       // As long as all tabs are loaded at the start, we don't need this check.
       // It is here to ensure similar behavior to stack
       if (!isInPreview) {
-        if (navigation.getState()?.type !== 'tab') {
-          throw new Error(
-            `Trigger component can only be used in the tab screen. Current route: ${route.name}`
-          );
-        }
+        // TODO(@ubax): `type` was removed from navigation state, so we can no longer assert that a
+        // Trigger is mounted inside a tab navigator. Rework to resolve navigator kind from the
+        // static layout config.
+        // if (navigation.getState()?.type !== 'tab') {
+        //   throw new Error(
+        //     `Trigger component can only be used in the tab screen. Current route: ${route.name}`
+        //   );
+        // }
         const options = convertTabPropsToOptions(props, true);
         navigation.setOptions(options);
       }